refactor: add greeting module
