+++ main.js
@@ -1,27 +1,12 @@
-import { throwError } from "./src/throw-error/index.js";
+// import { throwError } from "./src/throw-error/index.js";
 
-const USERNAME_REG_EXP = /^--username=('[^']'|"[^"]"|\S+)$/;
+import {
+  getPassedArgument,
+  getUserName,
+  showGreeting,
+} from "./src/greeting/index.js";
 
-const getPassedArgument = () => {
-  const passedArgument = process.argv[2];
-  if (!passedArgument) {
-    throwError();
-  }
-  return passedArgument;
-};
 
-const getUserName = (passedArgument) => {
-  const matched = passedArgument.match(USERNAME_REG_EXP);
-  if (!matched) {
-    throwError();
-  }
-  const [_, username] = matched;
-  return username;
-};
+const FAREWELL_MESSAGE = "Thank you for using File Manager, Username, goodbye!";
 
-const showGreeting = (username) => {
-   console.log(`Welcome to the File Manager, ${username}!`);
-}
-
-showGreeting(getUserName(getPassedArgument()))
-
+showGreeting(getUserName(getPassedArgument()));
